requirements.txt
python-telegram-bot==22.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.1
python-dotenv==1.1.0